requirements.txt
Django==2.0
generic==0.3.1
Pillow==5.0.0
pytz==2018.3
psycopg2
django-grappelli
django-ckeditor
